Style: handler 안 쓰는 핸들러 정리 및 삭제
